Update source files to define namespace
Was previously only defining the namespace in header files
and then definint every implementation as namespace::fn etc.
which I found to be annoying. So now defining the namespace
in both the header file and the source file.

No idea if this is a correct thing to do but at least it means
that I don't have to write '::' as much all the time.